fix warnings
use ints where possible instead of size_t
lengthen strings for snprintf to make sure content fits

+++ src/graphicsn.h
@@ -70,10 +70,10 @@ class graphicsN_c : public graphics_c {
     void setCursorMode(cursorMode_e mode);
     void setShowGrid(bool on);
     bool getShowGrid(void) const { return grid; }
-    size_t getCursorX(void) const { return cursorX; }
-    size_t getCursorY(void) const { return cursorY; }
-    size_t getCursorW(void) const { return cursorW; }
-    size_t getCursorH(void) const { return cursorH; }
+    int getCursorX(void) const { return cursorX; }
+    int getCursorY(void) const { return cursorY; }
+    int getCursorW(void) const { return cursorW; }
+    int getCursorH(void) const { return cursorH; }
     void setCursor(size_t x, size_t y, size_t w, size_t h);
 
     // set a status text, this text automatically vanishes after a while
@@ -107,7 +107,7 @@ class graphicsN_c : public graphics_c {
     bool editorMode;
     cursorMode_e cursorMode;
     bool grid;
-    size_t cursorX, cursorY, cursorW, cursorH;
+    int cursorX, cursorY, cursorW, cursorH;
     std::string statusText;
     uint16_t statusTime;
     const surface_c * overlay;
